:art: define and implement view class interface and move files
