+++ src/models/platform/message.ts
@@ -1,6 +1,12 @@
+import { GratitudeSummaryViewProps } from './slack/views/view-gratitude-summary';
+
 export type Message = string | View | InteractiveView
 
-export abstract class View {}
+export abstract class View {
+  static gratitudeSummary = (props: GratitudeSummaryViewProps) => {}
+}
 
 // TODO: add interface methods (gratitudeMessage, coffeeRoulette...)
-export abstract class InteractiveView {}
+export abstract class InteractiveView {
+  static gratitudeMessage = () => {}
+}
