+++ src/models/platform/slack/views/view-gratitude-summary.ts
@@ -1,7 +1,7 @@
-import { GratitudeSummaryMessage } from "../../../../models/database/gratitude-message";
-import { I18n } from "../../../i18n/i18n";
-import { getDateFormatted } from "../../../logger/logger";
-import { SlackView } from ".";
+import { GratitudeSummaryMessage } from "../../../database/gratitude-message";
+import { SlackView } from "./views";
+import { I18n } from "../../../../services/i18n/i18n";
+import { getDateFormatted } from "../../../../services/logger/logger";
 
 export interface GratitudeSummaryViewProps {
   image: string,
@@ -13,7 +13,6 @@ const toMessage = ({ users, createdAt, text }: GratitudeSummaryMessage) => {
   return `• *${users.map(({ id }) => `<@${id}>`).join(", ")}* \`${getDateFormatted(createdAt)}\`: ${text}`;
 }
 
-// TODO: maybe add tests
 export const GratitudeSummaryView = async ({ image, sent, received }: GratitudeSummaryViewProps): Promise<SlackView> => {
   const i18n = await I18n.getInstance()
   const hasSent = sent !== undefined && sent.length > 0
